save-analysis: move csv dumping stuff to its own module and rename
